+ Fixed a bug where the Help menu would appear on the second line of the menu at default size - this is still hard-coded, we should change it to be based on the dimensions of the bitmap loaded. + Fixed a bug where the Capture Window mode was broken. Also cleaned up the code in the function; it was pretty bad. + Added new dialog which is modeless and just displays a message for the user to click on a window or screen. I made it modeless because if I made it modal, the OnCaptureChanged function in RecorderView would not be called. + Cleaned up the menu for multiple monitor options. It now will show you different menu options depending on how many monitors you have connected. + Added code to OnCaptureChanged for recording a specific monitor. + Fixed a bunch of warnings. It now compiles without warnings on level 2! I would like to get that up to 3 or 4 but there were 200+ and I didn't want to spend the time on warnings at the moment. + Updated bugs file to reflect recent bug fixes, new bugs discovered and new bugs reported by Nick. + Added BasicMessage.cpp and BasicMessage.h

+++ CamStudio/Recorder/AVI.cpp
@@ -431,7 +431,7 @@ bool CamAVIFile::CompressionOpts(AVICOMPRESSOPTIONS& rOptions, CAVIStream& raviS
 		rOptions.dwInterleaveEvery = m_sAudioFormat.m_iInterleaveFactor;;
 		if (m_sAudioFormat.m_iInterleavePeriod == MILLISECONDS) {
 			double interfloat = (((double) m_sAudioFormat.m_iInterleaveFactor) * ((double) m_sVideoOpts.m_iFramesPerSecond))/1000.0;
-			rOptions.dwInterleaveEvery = (interfloat < 0.0) ? 1 : interfloat;
+			rOptions.dwInterleaveEvery = (DWORD)((interfloat < 0.0) ? 1 : interfloat);
 		}
 	} else {
 		ASSERT(bAudioStream);
@@ -448,7 +448,7 @@ bool CamAVIFile::CompressionOpts(AVICOMPRESSOPTIONS& rOptions, CAVIStream& raviS
 		rOptions.dwInterleaveEvery = m_sAudioFormat.m_iInterleaveFactor;;
 		if (m_sAudioFormat.m_iInterleavePeriod == MILLISECONDS) {
 			double interfloat = (((double) m_sAudioFormat.m_iInterleaveFactor) * ((double) m_sVideoOpts.m_iFramesPerSecond))/1000.0;
-			rOptions.dwInterleaveEvery = (interfloat < 0.0) ? 1 : interfloat;
+			rOptions.dwInterleaveEvery = (DWORD)((interfloat < 0.0) ? 1 : interfloat);
 		}
 		LONG lSize = 0;
 		raviStream.ReadFormat(raviStream.Start(), 0, &lSize);
@@ -656,7 +656,7 @@ bool CamAVIFile::FadeOut(const CString& strAVIIn, const CString& strAVIOut)
 		bResult = bResult && (0 != lpBitmap);
 		bResult = bResult && cxImage.Decode((LPBYTE)lpImage, lpBitmap->biSizeImage, CXIMAGE_FORMAT_UNKNOWN);
 		dBrightness -= dIncr;
-		long lBrightness = (MIN_BRIGHTNESS < dBrightness) ? dBrightness : MIN_BRIGHTNESS;
+		long lBrightness = (long)((MIN_BRIGHTNESS < dBrightness) ? dBrightness : MIN_BRIGHTNESS);
 		bResult = bResult && cxImage.Light(lBrightness);
 		hResult = bResult && aviCompressedStreamOut.Write(lPos, 1L, (LPBYTE)cxImage.GetDIB() + lOffsetImage, lpBitmap->biSizeImage, 0, 0, 0);
 		CAVI::OnError(hResult);
